requirements.txt
pyannote.audio
torch
torchaudio
pydub